Extract CtRegionMap Flame host out of lib/widgets (Refs #1878)
Move FlameGame tap/camera logic to features/game/flame/ct_region_map_game.dart
and GameWidget embedding to chrome/region_map_game_viewport.dart so
ct_region_map.dart no longer depends on a Flame re-export shim.

Tighten repo.no_flame_in_widgets: match import/export with single or double
quotes; skip whole-line // comments. Extend SPEC/program/repo-lint.md and
add regression tests for the checker.

Made-with: Cursor
