+++ app/test/ct_region_map_widget_test.dart
@@ -34,6 +34,7 @@ import 'package:colonizethis_app/features/game/flame/civilian_icon_cache.dart';
 import 'package:colonizethis_app/features/game/flame/province_label_icon_cache.dart';
 import 'package:colonizethis_app/features/game/flame/terrain_tileset.dart';
 import 'package:colonizethis_app/features/game/flame/town_icon_cache.dart';
+import 'package:colonizethis_app/features/game/flame/ct_region_map_game.dart';
 import 'package:colonizethis_app/features/game/flame/transport_overlay_tileset.dart';
 import 'package:colonizethis_app/widgets/ct_region_map.dart'
     show BaseLayerDisplayMode, CtRegionMap, CtMapVisibilityMode;
@@ -46,8 +47,8 @@ CtRegionMapComponent ctRegionMapComponentFromTester(WidgetTester tester) {
   );
   expect(finder, findsOneWidget);
   final gameWidget = tester.widget(finder);
-  final game = (gameWidget as dynamic).game;
-  return (game as dynamic).debugMapComponentForTest as CtRegionMapComponent;
+  final game = (gameWidget as dynamic).game as CtRegionMapGame;
+  return game.debugMapComponentForTest;
 }
 
 void main() {
